refactor(ESLint): ♻️  Replace config from `svelte3` to `svelte`
